Extracted Functions to read image files into a utility header.
Read Fiji Montage is working now.
