+++ src/Plugins/ITKImageProcessing/src/ITKImageProcessing/Common/ReadImageUtils.cpp
@@ -1,0 +1,60 @@
+#include "ReadImageUtils.hpp"
+
+namespace cxItkImageReader
+{
+
+//------------------------------------------------------------------------------
+Result<OutputActions> ReadImagePreflight(const std::string& fileName, DataPath imageGeomPath, std::string cellDataName, DataPath arrayPath)
+{
+  OutputActions actions;
+
+  try
+  {
+    itk::ImageIOBase::Pointer imageIO = itk::ImageIOFactory::CreateImageIO(fileName.c_str(), itk::ImageIOFactory::ReadMode);
+    if(imageIO == nullptr)
+    {
+      return MakeErrorResult<OutputActions>(-5, fmt::format("ITK could not read the given file \"{}\". Format is likely unsupported.", fileName));
+    }
+
+    imageIO->SetFileName(fileName);
+    imageIO->ReadImageInformation();
+
+    itk::ImageIOBase::IOComponentEnum component = imageIO->GetComponentType();
+
+    std::optional<DataType> numericType = ITK::ConvertIOComponentToDataType(component);
+    if(!numericType.has_value())
+    {
+      return MakeErrorResult<OutputActions>(-4, fmt::format("Unsupported pixel component: {}", imageIO->GetComponentTypeAsString(component)));
+    }
+
+    uint32 nDims = imageIO->GetNumberOfDimensions();
+
+    std::vector<size_t> dims = {1, 1, 1};
+    std::vector<float32> origin = {0.0f, 0.0f, 0.0f};
+    std::vector<float32> spacing = {1.0f, 1.0f, 1.0f};
+
+    for(uint32 i = 0; i < nDims; i++)
+    {
+      dims[i] = static_cast<usize>(imageIO->GetDimensions(i));
+      origin[i] = static_cast<float32>(imageIO->GetOrigin(i));
+      spacing[i] = static_cast<float32>(imageIO->GetSpacing(i));
+    }
+
+    uint32 nComponents = imageIO->GetNumberOfComponents();
+
+    // DataArray dimensions are stored slowest to fastest, the opposite of ImageGeometry
+    std::vector<usize> arrayDims(dims.crbegin(), dims.crend());
+
+    std::vector<usize> cDims = {nComponents};
+
+    actions.appendAction(std::make_unique<CreateImageGeometryAction>(std::move(imageGeomPath), std::move(dims), std::move(origin), std::move(spacing), cellDataName));
+    actions.appendAction(std::make_unique<CreateArrayAction>(*numericType, std::move(arrayDims), std::move(cDims), std::move(arrayPath)));
+
+  } catch(const itk::ExceptionObject& err)
+  {
+    return MakeErrorResult<OutputActions>(-55557, fmt::format("ITK exception was thrown while processing input file: {}", err.what()));
+  }
+
+  return {std::move(actions)};
+}
+} // namespace cxItkImageReader
